feat: Add skip_before_action to ActivitiesController
This commit adds the `skip_before_action` method to the `ActivitiesController` in order to bypass authenticity token verification. This change allows for the creation of a new activity without requiring an authenticity token.

The commit also includes changes to the routes file, adding the `create` action to the list of allowed actions for the activities resource.

Additionally, this commit updates the developers.md document with some notes about future improvements and plans for the API.